week-10 implementations are in progress, website re-designed(not completed)

+++ src/main/java/ih/ifbs/domain/Airline.java
@@ -62,7 +62,7 @@ public class Airline extends EntityClass {
         return fleetSize;
     }
 
-    public double getTotalDestinations() {
+    public int getTotalDestinations() {
         return totalDestinations;
     }
